Load and Execute Script
This is not an executable so linking with the lower level drivers may need to be done manually because I'm not going to bother with the project structure here.  The dir for the drivers is flat.  This item is one dir down.  This is a user interface and is built using the lower level drivers. The file or script is as such.  The left hand column contains the axis and the right contains the steps.  It is normal for this to be a min ratio of the image desired because the interface allows for this value to be scaled.  The distance the axis travels is based on the axis screw.  The following script draws Sponge Bob holding a drink and is used with a metals etcher but I have also seen a pen and note pad used:

X	2
Y	3
Z	DOWN
Y	5
X	3
Y	4
X	1
Y	3
X	2
Z	UP
X	6
Y	6
Z	DOWN
X	1
Y	7
X	2
Y	8
X	3
Y	9
X	2
Y	10
X	3.5
X	2
Y	11
X	3.5
X	2
Y	12
X	4
Y	13
X	6
Y	14
X	1
Y	13
X	3
Y	12
X	5
Y	11
X	6
Y	10
X	5
Y	9
X	3
X	4
Y	8
X	5
Y	7
X	6
Y	6
Z	UP
Y	9
Z	DOWN
Y	12
X